Remove Turbo.js and fix category selection on the edit page
