+++ app/views/recipes/index.html.erb
@@ -1,2 +1,2 @@
 <h1>Recipes#index</h1>
-<p>Find me in app/views/recipes/index.html.erb</p>
+<%= render 'recipes/recipes' %>
