+++ config/webpack/environment.js
@@ -1,3 +1,5 @@
-const { environment } = require('@rails/webpacker')
+const { environment } = require('@rails/webpacker');
 
-module.exports = environment
+environment.loaders.delete('nodeModules');
+
+module.exports = environment;
